Clean and document code
