+++ tests-e2e/cypress/support/index.ts
@@ -4,4 +4,5 @@
 import 'cypress-wait-until';
 import './server_api_commands.ts';
 import './ui_commands.ts';
-
+import './verification_commands.ts';
+import './navigation_commands.ts';
