+++ tests-e2e/cypress/support/server_api_commands.ts
@@ -9,11 +9,6 @@ import {Post} from 'mattermost-redux/types/posts';
 import users from '../fixtures/users';
 import {httpStatusOk, httpStatusCreated} from '../support/constants';
 
-// *****************************************************************************
-// Authentication
-// https://api.mattermost.com/#tag/authentication
-// *****************************************************************************
-
 function apiLogin(username = 'user-1', password : string | null = null) : Cypress.Chainable<Cypress.Response> {
     return cy.request({
         headers: {'X-Requested-With': 'XMLHttpRequest'},
@@ -30,6 +25,8 @@ function apiLogin(username = 'user-1', password : string | null = null) : Cypres
 }
 Cypress.Commands.add('apiLogin', apiLogin);
 
+// Return a function to create either a public or private channel depending on
+// the channelType parameter.
 function apiCreateChannel(channelType: ChannelType) : ((teamId: string, name: string, displayName: string) => Cypress.Chainable<Channel>) {
     return (teamId: string, name: string, displayName: string): Cypress.Chainable<Channel> => {
         return cy.request({
@@ -52,6 +49,36 @@ function apiCreateChannel(channelType: ChannelType) : ((teamId: string, name: st
 }
 Cypress.Commands.add('apiCreatePublicChannel', apiCreateChannel('O'));
 Cypress.Commands.add('apiCreatePrivateChannel', apiCreateChannel('P'));
+
+function apiCreateGroupMessage(userIds : string[]) : Cypress.Chainable<Channel> {
+    return cy.request({
+        headers: {'X-Requested-With': 'XMLHttpRequest'},
+        url: '/api/v4/channels/group',
+        method: 'POST',
+        body: userIds,
+    }).then((response: Cypress.Response) => {
+        expect(response.status).to.equal(httpStatusCreated);
+
+        const channel = response.body as Channel;
+        return cy.wrap(channel);
+    });
+}
+Cypress.Commands.add('apiCreateGroupMessage', apiCreateGroupMessage);
+
+function apiCreateDirectMessage(selfId : string, otherId : string) : Cypress.Chainable<Channel> {
+    return cy.request({
+        headers: {'X-Requested-With': 'XMLHttpRequest'},
+        url: '/api/v4/channels/direct',
+        method: 'POST',
+        body: [selfId, otherId],
+    }).then((response: Cypress.Response) => {
+        expect(response.status).to.equal(httpStatusCreated);
+
+        const channel = response.body as Channel;
+        return cy.wrap(channel);
+    });
+}
+Cypress.Commands.add('apiCreateDirectMessage', apiCreateDirectMessage);
 
 function apiGetTeamByName(name: string) : Cypress.Chainable<Team> {
     return cy.request({
@@ -82,6 +109,21 @@ function apiGetUserByUsername(name: string) : Cypress.Chainable<UserProfile> {
 }
 Cypress.Commands.add('apiGetUserByUsername', apiGetUserByUsername);
 
+function apiGetUsers(usernames : string[] = []) : Cypress.Chainable<UserProfile[]> {
+    return cy.request({
+        headers: {'X-Requested-With': 'XMLHttpRequest'},
+        url: '/api/v4/users/usernames',
+        method: 'POST',
+        body: usernames,
+    }).then((response: Cypress.Response) => {
+        expect(response.status).to.equal(httpStatusOk);
+
+        const userList = response.body as UserProfile[];
+        return cy.wrap(userList);
+    });
+}
+Cypress.Commands.add('apiGetUsers', apiGetUsers);
+
 function apiCreatePost(channelId: string, message: string, fileIds: string[] = []) : Cypress.Chainable<Post> {
     return cy.request({
         headers: {'X-Requested-With': 'XMLHttpRequest'},
@@ -101,13 +143,6 @@ function apiCreatePost(channelId: string, message: string, fileIds: string[] = [
 }
 Cypress.Commands.add('apiCreatePost', apiCreatePost);
 
-function apiCreateMultiplePosts(channelId: string, numMessages: number) : void {
-    for (let i = 0; i < numMessages; i++) {
-        cy.apiCreatePost(channelId, 'lorem ipsum ' + i);
-    }
-}
-Cypress.Commands.add('apiCreateMultiplePosts', apiCreateMultiplePosts);
-
 function apiGetChannelByName(teamName: string, channelName: string) : Cypress.Chainable<Channel> {
     return cy.request({
         headers: {'X-Requested-With': 'XMLHttpRequest'},
@@ -122,6 +157,19 @@ function apiGetChannelByName(teamName: string, channelName: string) : Cypress.Ch
     });
 }
 Cypress.Commands.add('apiGetChannelByName', apiGetChannelByName);
+
+function apiMakeChannelReadOnly(channelId: string) : Cypress.Chainable<Cypress.Response> {
+    return cy.request({
+        headers: {'X-Requested-With': 'XMLHttpRequest'},
+        url: `/api/v4/channels/${channelId}/moderations/patch`,
+        method: 'PUT',
+        body: [{name: 'create_post', roles: {members: false, guests: false}}],
+    }).then((response: Cypress.Response) => {
+        expect(response.status).to.equal(httpStatusOk);
+        return cy.wrap(response);
+    });
+}
+Cypress.Commands.add('apiMakeChannelReadOnly', apiMakeChannelReadOnly);
 
 function apiExportChannel(channelId: string, expectedStatus: number = httpStatusOk) : Cypress.Chainable<string> {
     const endpoint = '/plugins/com.mattermost.plugin-channel-export/api/v1/export';
@@ -140,69 +188,3 @@ function apiExportChannel(channelId: string, expectedStatus: number = httpStatus
     });
 }
 Cypress.Commands.add('apiExportChannel', apiExportChannel);
-
-function apiMakeChannelReadOnly(channelId: string) : Cypress.Chainable<Cypress.Response> {
-    return cy.request({
-        headers: {'X-Requested-With': 'XMLHttpRequest'},
-        url: `/api/v4/channels/${channelId}/moderations/patch`,
-        method: 'PUT',
-        body: [{name: 'create_post', roles: {members: false, guests: false}}],
-    }).then((response: Cypress.Response) => {
-        expect(response.status).to.equal(httpStatusOk);
-        return cy.wrap(response);
-    });
-}
-Cypress.Commands.add('apiMakeChannelReadOnly', apiMakeChannelReadOnly);
-
-/**
- * Creates a group channel directly via API
- * This API assume that the user is logged in and has cookie to access
- * @param {String} userIds - IDs of users as member of the group
- * All parameters required except purpose and header
- */
-function apiCreateGroupMessage(userIds : string[]) : Cypress.Chainable<Channel> {
-    return cy.request({
-        headers: {'X-Requested-With': 'XMLHttpRequest'},
-        url: '/api/v4/channels/group',
-        method: 'POST',
-        body: userIds,
-    }).then((response: Cypress.Response) => {
-        expect(response.status).to.equal(httpStatusCreated);
-
-        const channel = response.body as Channel;
-        return cy.wrap(channel);
-    });
-}
-Cypress.Commands.add('apiCreateGroupMessage', apiCreateGroupMessage);
-
-function apiCreateDirectMessage(userIds : string[]) : Cypress.Chainable<Channel> {
-    return cy.request({
-        headers: {'X-Requested-With': 'XMLHttpRequest'},
-        url: '/api/v4/channels/direct',
-        method: 'POST',
-        body: userIds,
-    }).then((response: Cypress.Response) => {
-        expect(response.status).to.equal(httpStatusCreated);
-
-        const channel = response.body as Channel;
-        return cy.wrap(channel);
-    });
-}
-Cypress.Commands.add('apiCreateDirectMessage', apiCreateDirectMessage);
-
-/**
-* Gets users
-*/
-Cypress.Commands.add('apiGetUsers', (usernames : string[] = []) => {
-    return cy.request({
-        headers: {'X-Requested-With': 'XMLHttpRequest'},
-        url: '/api/v4/users/usernames',
-        method: 'POST',
-        body: usernames,
-    }).then((response: Cypress.Response) => {
-        expect(response.status).to.equal(httpStatusOk);
-
-        const userList = response.body as UserProfile[];
-        return cy.wrap(userList);
-    });
-});
